pyproject.toml
[tool.poetry]
name = "forest-analyzer"
version = "0.1.0"
description = "Textual forest data analyzer"
authors = ["ronikeuru <roni@three.consulting>"]
readme = "README.md"
packages = [{include = "forest_analyzer"}]

[tool.poetry.dependencies]
python = "^3.10"
python-dotenv = "^1.0.0"
psycopg2-binary = "^2.9.6"
geoalchemy2 = "^0.13.1"
llama-index = "^0.5.9"
langchain = "^0.0.134"
openai = "^0.27.4"
numpy = "^1.24.2"
geopandas = "^0.12.2"
altair = "^4.2.2"
branca = "^0.6.0"
jinja2 = "^3.1.2"
requests = "^2.28.2"
folium = "^0.14.0"
plotly = "^5.14.1"
streamlit = "^1.21.0"
streamlit-folium = "^0.11.1"
wikipedia = "^1.4.0"
nbformat = "^5.8.0"
chart-studio = "^1.1.0"
shapely = "^2.0.1"
geoplot = "^0.5.1"
pydeck = {extras = ["jupyter"], version = "^0.8.0"}
pydantic = "^1.10.7"

[tool.poetry.group.dev.dependencies]
ipykernel = "^6.22.0"
spatialite = "^0.0.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
